refactor: move to do card to folder cards
